Enhance AOT-Support (Issue#3171)
replace registerBeanDefinition callback with a factoryMethod
split out a RouterFunctionHolderFactory from the Registrar class and provide its bean in the AutoConfiguration
set refresh scope on the RouterFunctionHolder bean only if RefreshScope bean in context and add warning
